Remove sanitize_url from munge script
This will currently cause issues: the fold-mappings won't quite work as
some new urls will need some sanitisation and we are currently not doing
any.

We need to tease apart what we're doing when we c14n a url - query
string removal isn't always approprate for new urls, but we do need to
do some work on them.

Pushed for review next week with the team.